Add files via upload
It's a Task 1 Number Game 
In this Task we have to write a code to generate a random number between 1 to 100 and user has to guess it using cosole. User has given 3 attempts to guess the correct number and if the user guesses correct number the score is added by 1. And if the use losses in total 3 attempts user has given 1 option to play the game or not.if user write yes the game continues and if user writes no on console the game will end.