Add files via upload
Its a 2nd The name of the task is  "Student Grade Calculator". In this task we have to take input of marks of user in each subject out of 100 and then we have to calculate the total marks , calculate average percentage, Grade calculation and show the result as above operation occurs.